MBX-2530 Move Inapp model
